install/requirements.txt
rich == 12.4.0
click == 8.1.3
nipype==1.6.1
nilearn==0.9.1
pybids == 0.15.1
pyyaml == 6.0
numpy == 1.22.3 
